pyproject.toml
[tool.poetry]
name = "members"
version = "0.1.0"
description = "Members management"
authors = ["Konstantinos Vitiniotis <kostasvit@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
django-environ = "^0.11.2"
django-filter = "^24.2"
django-widget-tweaks = "^1.5.0"
django-import-export = "^4.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
